Rearrange the code to enable future implementation of SimProducer class
