+++ Validation/HcalHits/plugins/BuildFile.xml
@@ -1,5 +1,6 @@
 <use name="CommonTools/UtilAlgos"/>
 <use name="DataFormats/HcalDetId"/>
+<use name="DataFormats/HepMCCandidate"/>
 <use name="DQMServices/Core"/>
 <use name="FWCore/Framework"/>
 <use name="FWCore/ParameterSet"/>
